Move word2vec tests to the word2vec module
